Add code (disabled) for both setting ssl/tls and initial password
When creating a user windows aduc prompts you to set the initial
password.  Currently this fails with and exception
Maybe related to the above some microsoft documentation seems to
indicate at least older versions of windows expected ssl/tls to
be enabled in order write that attribute. Note: even though we
have done a sasl bind we don't seem to be writing encrypted stuff
over ldap.